pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Artur Medvedev <varouser@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.4"
django-filter = "^24.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
